# libs/client/rest/src/nutri-openapi.yaml
openapi: 3.0.0
paths:
...
  /api/recipes/{id}:
    get:
      operationId: RecipesController_findOne
      summary: Find a recipe by ID
      parameters:
        - name: id
          required: true
          in: path
          description: Recipe ID
          schema:
            type: string
      responses:
        '200':
          description: Recipe found
        '404':
          description: Recipe not found
      tags:
        - Recipe Management
    delete:
      operationId: RecipesController_remove
      summary: Delete a recipe
      parameters:
        - name: id
          required: true
          in: path
          description: Recipe ID
          schema:
            type: string
      responses:
        '200':
          description: Recipe successfully deleted
        '404':
          description: Recipe not found
      tags:
        - Recipe Management
...
  /api/schemas/recipe/{operation}:
    get:
      operationId: RecipeSchemasController_getRecipeSchema
      summary: Get recipe JSON schema for specific operation
      parameters:
        - name: operation
          required: true
          in: path
          description: Operation type (create, update, etc.)
          schema:
            type: string
      responses:
        '200':
          description: Schema returned successfully
      tags:
        - Recipe Schemas
  /api/schemas/recipe-ingredient/{operation}:
    get:
      operationId: RecipeSchemasController_getIngredientSchema
      summary: Get recipe ingredient JSON schema for specific operation
      parameters:
        - name: operation
          required: true
          in: path
          description: Operation type (create, update, etc.)
          schema:
            type: string
      responses:
        '200':
          description: Schema returned successfully
      tags:
        - Recipe Schemas
  /api/schemas/recipe-instruction/{operation}:
    get:
      operationId: RecipeSchemasController_getInstructionSchema
      summary: Get recipe instruction JSON schema for specific operation
      parameters:
        - name: operation
          required: true
          in: path
          description: Operation type (create, update, etc.)
          schema:
            type: string
      responses:
        '200':
          description: Schema returned successfully
      tags:
        - Recipe Schemas
  /api/recipes/{recipeId}/ingredients:
    post:
      operationId: RecipeIngredientsController_create
      summary: Create a recipe ingredient
      parameters:
        - name: recipeId
          required: true
          in: path
          description: Recipe ID
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateRecipeIngredientDto'
      responses:
        '201':
          description: Ingredient created successfully
        '400':
          description: Bad request
      tags:
        - Recipe Ingredients
    get:
      operationId: RecipeIngredientsController_findAll
      summary: Find all ingredients for a recipe
      parameters:
        - name: recipeId
          required: true
          in: path
          description: Recipe ID
          schema:
            type: string
      responses:
        '200':
          description: List of ingredients returned
        '404':
          description: Recipe not found
      tags:
        - Recipe Ingredients
  /api/recipes/{recipeId}/ingredients/batch:
    post:
      operationId: RecipeIngredientsController_createBatch
      summary: Create multiple recipe ingredients
      parameters:
        - name: recipeId
          required: true
          in: path
          description: Recipe ID
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: '#/components/schemas/CreateRecipeIngredientDto'
      responses:
        '201':
          description: Ingredients created successfully
        '400':
          description: Bad request
      tags:
        - Recipe Ingredients
  /api/recipes/{recipeId}/instructions/batch:
    post:
      operationId: RecipeInstructionsController_createBatch
      summary: Create multiple recipe instructions
      parameters:
        - name: recipeId
          required: true
          in: path
          description: Recipe ID
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: '#/components/schemas/CreateRecipeInstructionDto'
      responses:
        '201':
          description: Instructions created successfully
        '400':
          description: Bad request
      tags:
        - Recipe Instructions
  /api/recipes/{recipeId}/instructions:
    get:
      operationId: RecipeInstructionsController_findAll
      summary: Find all instructions for a recipe
      parameters:
        - name: recipeId
          required: true
          in: path
          description: Recipe ID
          schema:
            type: string
      responses:
        '200':
          description: List of instructions returned
        '404':
          description: Recipe not found
      tags:
        - Recipe Instructions
...
  /api/recipes/drafts/{id}:
    get:
      operationId: RecipeDraftsController_getDraft
      summary: Get a recipe draft by ID
      parameters:
        - name: id
          required: true
          in: path
          description: Draft ID
          schema:
            type: string
      responses:
        '200':
          description: Draft found
        '404':
          description: Draft not found
      tags:
        - Recipe Drafts
    put:
      operationId: RecipeDraftsController_updateDraftBasic
      summary: Update basic recipe draft information
      parameters:
        - name: id
          required: true
          in: path
          description: Draft ID
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateRecipeDto'
      responses:
        '200':
          description: Draft updated successfully
        '404':
          description: Draft not found
      tags:
        - Recipe Drafts
    delete:
      operationId: RecipeDraftsController_deleteDraft
      summary: Delete a recipe draft
      parameters:
        - name: id
          required: true
          in: path
          description: Draft ID
          schema:
            type: string
      responses:
        '200':
          description: Draft deleted successfully
        '404':
          description: Draft not found
      tags:
        - Recipe Drafts
  /api/recipes/drafts/{id}/ingredients:
    put:
      operationId: RecipeDraftsController_updateDraftIngredients
      summary: Update ingredients in a recipe draft
      parameters:
        - name: id
          required: true
          in: path
          description: Draft ID
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: '#/components/schemas/CreateRecipeIngredientDto'
      responses:
        '200':
          description: Draft ingredients updated successfully
        '404':
          description: Draft not found
      tags:
        - Recipe Drafts
  /api/recipes/drafts/{id}/steps:
    put:
      operationId: RecipeDraftsController_updateDraftInstructions
      summary: Update instructions in a recipe draft
      parameters:
        - name: id
          required: true
          in: path
          description: Draft ID
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: '#/components/schemas/CreateRecipeInstructionDto'
      responses:
        '200':
          description: Draft instructions updated successfully
        '404':
          description: Draft not found
      tags:
        - Recipe Drafts
  /api/recipes/drafts/{id}/finalize:
    post:
      operationId: RecipeDraftsController_finalizeDraft
      summary: Finalize a recipe draft to create a complete recipe
      parameters:
        - name: id
          required: true
          in: path
          description: Draft ID
          schema:
            type: string
      responses:
        '201':
          description: Draft finalized successfully
        '404':
          description: Draft not found
      tags:
        - Recipe Drafts
info:
  title: Nutri API
  description: Internal REST API documentation
  version: 0.0.1
  contact: {}
...
servers: []
components:
  schemas:
...
    CreateRecipeDto:
      type: object
      properties:
        title:
          type: string
          minLength: 3
          maxLength: 100
        description:
          type: string
        sourceId:
          type: string
        sourceUrl:
          type: string
        cookingTime:
          type: string
        prepTime:
          type: string
        servingsText:
          type: string
        servingsMin:
          type: number
          minimum: 1
        servingsMax:
          type: number
          minimum: 1
        tagIds:
          type: array
          items:
            type: string
      required:
        - title
    CreateRecipeIngredientDto:
      type: object
      properties:
        recipeId:
          type: string
        partId:
          type: string
        quantity:
          type: number
          minimum: 0
        quantityText:
          type: string
        minQuantity:
          type: number
          minimum: 0
        maxQuantity:
          type: number
          minimum: 0
        unit:
          type: string
        unitText:
          type: string
        ingredientText:
          type: string
        extraInfo:
          type: string
        foodId:
          type: string
      required:
        - recipeId
        - quantity
        - unit
        - ingredientText
    CreateRecipeInstructionDto:
      type: object
      properties:
        recipeId:
          type: string
        content:
          type: string
        type:
          type: object
        order:
          type: number
          minimum: 1
        listIndex:
          type: number
          minimum: 1
      required:
        - recipeId
        - content
        - type
        - order